Removed dropdown select styles and moved Meteor startup's uploadServer to sarai-images-data package
